Make sure zram swap has highest priority
On systems with that use a swap partition or file on disk, the system may use the fstab to enable the disk-based swap first, before systemd starts this service. This will result in the system not using the zram swap, as by default, swap spaces are used in the order in which they were enabled. Setting priority to the maximum of 32767 avoids this problem.